Added Delete command. Rearranged command package
