+++ src/main/java/net/szum123321/textile_backup/commands/manage/ListBackupsCommand.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.szum123321.textile_backup.commands.restore;
+package net.szum123321.textile_backup.commands.manage;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import net.minecraft.server.command.CommandManager;
